analyzer/requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
gspread>=6.0.0
google-auth>=2.27.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
pandas>=2.0.0
pyhwp>=0.1b12
